move search to own component
